Backend/requirements.txt

Flask==3.0.0
flask-cors==4.0.0
tensorflow==2.16.2
opencv-python==4.8.1.78
numpy==1.26.4
Pillow==10.1.0
werkzeug==3.0.1
python-dotenv==1.0.0

